feat(web): split root layout into marketing and app layouts
