+++ apps/web/src/app/app/layout.tsx
@@ -1,0 +1,15 @@
+import type { ReactNode } from 'react';
+import { Toaster } from '@repo/ui/components/sonner';
+
+type AppLayoutProps = Readonly<{
+  children: ReactNode;
+}>;
+
+export default async function AppLayout({ children }: AppLayoutProps) {
+  return (
+    <>
+      <>{children}</>
+      <Toaster />
+    </>
+  );
+}
